refactor functions and address comments
